fix: deleting first version of introduction
