+++ src/hooks/useManageRoute.ts
@@ -12,7 +12,7 @@ const useManageRoute = () => {
       if (loading) return;
 
       if (!user) {
-        resetTo(navigation, "DiscoverApp");
+        resetTo(navigation, "Introduction");
       } else {
         // await logInRevenueCat(user.id, user.email);
         resetTo(navigation, "HomeStack");
